+++ iBXTest/iBXTest/Sources/Controllers/NavigationTestController/NavigationBackgroundController.h
@@ -1,0 +1,13 @@
+//
+//  NavigationBackgroundController.h
+//  iBXTest
+//
+//  Created by Sergey Balalaev on 04/06/2017.
+//  Copyright © 2017 ByteriX. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface NavigationBackgroundController : UIViewController
+
+@end